+++ R/pca.R
@@ -2,7 +2,15 @@
 #' Get the eigenvalues of PCA
 #'
 #' @param dat The input data set.
+#' @param object An R object
+#' @param cutoffs Cutoffs for percent variance explained.
+#' @param ... Other arguments.
 #' @return An object of class "dq_pca": a numeric vector containing eigenvalues.
+#' @name dq_pca
+NULL
+#> NULL
+
+#' @rdname dq_pca
 #' @export
 dq_pca <- function(dat)
 {
@@ -18,3 +26,13 @@ dq_pca <- function(dat)
 }
 
 
+#' @rdname dq_pca
+#' @export
+summary.dq_pca <- function(object, cutoffs = c(0.95, 0.975, 0.99), ...)
+{
+  eig <- cumsum(object)/sum(object)
+  out <- lapply(cutoffs, function(cutoff) {
+    min(which(cutoff <= eig))
+  })
+  stats::setNames(as.data.frame(out), paste0(cutoffs*100, "%"))
+}